Simplify DcGroup
